Rename some files and move them into subfolder, remove unsed folder 'plots' from assest
