+++ src/components/plots/Heatmap.js
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react';
 import Plot from 'react-plotly.js';
-import { downloadSVG } from '../../utils/downLoadSvg';
+import { downloadSVG } from '../../utils/downloadHelpers/downLoadSvg';
 import orgMeta from '../../utils/organismMetadata.js';
 import { Popover, Button } from 'antd';
