+++ src/utils/chatHelper.js/addRemoveHandler.js
