+++ src/utils/chatHelper.js/plotConversion.js
