+++ src/utils/chatSideEffects.js
@@ -1,9 +1,9 @@
 import atlasapprox from "@fabilab/atlasapprox";
-import { buildAPIParams, buildAnswer } from "./nlpHelpers.js";
-import { handleNoApiIntents } from "./hanleNoApiIntent.js";
-import { handlePlotConversion } from "./plotConversion.js";
-import { handleAddRemove } from "./addRemoveHandler.js";
-import { handleErrors } from "./errorHandler.js"; 
+import { buildAPIParams, buildAnswer } from "./nlpResponseGenerator.js";
+import { handleNoApiIntents } from "./chatHelper.js/hanleNoApiIntent.js";
+import { handlePlotConversion } from "./chatHelper.js/plotConversion.js";
+import { handleAddRemove } from "./chatHelper.js/addRemoveHandler.js";
+import { handleErrors } from "./chatHelper.js/errorHandler.js"; 
 
 // updatePlotIntents: An array of intents that trigger a plot update.
 // These intents require either fetching data from the API or from previous plot state, and updating the plot accordingly.
