+++ src/utils/downloadHelpers/downloadFasta.js
