+++ src/utils/nlpResponseGenerator.js
